src/main/java/site/telion/skyprocoursework02/service/JavaQuestionService.java
package site.telion.skyprocoursework02.service;

import org.springframework.stereotype.Service;
import site.telion.skyprocoursework02.model.Question;

import java.util.*;

@Service
public class JavaQuestionService implements QuestionService {
    private final Set<Question> questionSet = new HashSet<>();

    @Override
    public Question add(String question, String answer) {
        Question newQuestion = new Question(question, answer);
        questionSet.add(newQuestion);
        return newQuestion;
    }

    @Override
    public Question add(Question question) {
        questionSet.add(question);
        return question;
    }

    @Override
    public Question remove(Question question) {
        questionSet.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return questionSet;
    }

    @Override
    public Question getRandomQuestion() {
        Random random = new Random();
        return new ArrayList<>(questionSet).get(random.nextInt(questionSet.size()));
    }
}
